+++ pytorch_lightning/trainer/trainer.py
@@ -2142,6 +2142,7 @@ class Trainer(
 
     @property
     def use_amp(self) -> bool:
+        rank_zero_deprecation("`Trainer.use_amp` is deprecated in v1.6.0 and will be removed in v1.8.0.")
         return self.precision == 16
 
     @property